# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Promptsheon API
  description: Version Control System for AI agent intelligence
  version: "1.0.0"
  contact:
    name: Promptsheon
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API key prefixed with ps_. Send as: Authorization: Bearer ps_..."
  schemas:
...
    Variable:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        required:
          type: boolean
        default:
          type: string
        description:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, writer, reader]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    APIKey:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
        last_used:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        revoked:
          type: boolean
paths:
...
  /api/v1/apikeys:
    post:
      summary: Create API key
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, name, role]
              properties:
                user_id:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [admin, writer, reader]
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Key created (key shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  record:
                    $ref: "#/components/schemas/APIKey"
    get:
      summary: List API keys for a user
      tags: [Auth]
      security: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
  /api/v1/apikeys/{id}:
    delete:
      summary: Revoke API key
      tags: [Auth]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Revoked
  /api/v1/auth/{provider}/login:
    get:
      summary: Initiate OAuth login
      tags: [Auth]
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
      responses:
        "302":
          description: Redirect to OAuth provider
  /api/v1/auth/{provider}/callback:
    get:
      summary: OAuth callback
      tags: [Auth]
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  key:
                    type: string
  /api/v1/users:
    get:
      summary: List users
      tags: [Users]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [admin, writer, reader]
                  default: reader
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/users/{id}:
    get:
      summary: Get user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: Delete user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/v1/prompts:
    get:
      summary: List prompts
      tags: [Prompts]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Prompt list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
    post:
      summary: Create prompt
      tags: [Prompts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name:
                  type: string
                description:
                  type: string
                content:
                  type: string
                variables:
                  type: array
                  items:
                    $ref: "#/components/schemas/Variable"
                tags:
                  type: array
                  items:
                    type: string
                model_hint:
                  type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /api/v1/prompts/{id}:
    get:
      summary: Get prompt
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
    put:
      summary: Update prompt
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                content:
                  type: string
                variables:
                  type: array
                  items:
                    $ref: "#/components/schemas/Variable"
                tags:
                  type: array
                  items:
                    type: string
                model_hint:
                  type: string
                status:
                  type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
    delete:
      summary: Delete prompt
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/v1/prompts/{id}/deploy:
    post:
      summary: Deploy an approved prompt
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /api/v1/prompts/{id}/archive:
    post:
      summary: Archive a prompt
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /api/v1/prompts/{id}/run:
    post:
      summary: Run a prompt via LLM
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  additionalProperties:
                    type: string
                provider:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: LLM response
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  model:
                    type: string
                  usage:
                    $ref: "#/components/schemas/Usage"
                  latency_ms:
                    type: integer
  /api/v1/prompts/{id}/preview:
    post:
      summary: Preview a prompt without executing
      tags: [Prompts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: Preview result
          content:
            application/json:
              schema:
                type: object
                properties:
                  rendered:
                    type: string
                    description: Prompt with variables substituted
                  estimated_tokens:
                    type: integer
                  estimated_cost:
                    type: number
                    format: double
                  variables_used:
                    type: object
                    additionalProperties:
                      type: string
...
  /api/v1/agents/{id}/export:
    get:
      summary: Export agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json, yaml]
      responses:
        "200":
          description: Agent in requested format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
            application/x-yaml:
              schema:
                type: string
...